Refactor Components and Actions
- in components:
- create dirs named Misc, Pages, Rails, Forms, and Modals
- classify components to dirs
- update rel paths for all components in code
- in actions/index.js:
- alphabetize const string exports and action functions
