+++ certain-death/src/actions/index.js
@@ -1,19 +1,19 @@
-import { v4 } from 'uuid';
 import axios from 'axios';
+import { v4 } from 'uuid';
 
 export const ADD_NOTE = 'ADD_NOTE';
-export const UPDATE_NOTE = 'UPDATE_NOTE';
 export const DELETE_NOTE = 'DELERE_NOTE';
-export const TOGGLE_DELETE = 'TOGGLE_DELETE';
-export const TITLE_SORT = 'TITLE_SORT';
+export const ERROR = 'ERROR';
+export const LOGIN = 'LOGIN';
 export const NEWEST_SORT = 'NEWEST_SORT';
 export const OLDEST_SORT = 'OLDEST_SORT';
-export const UPDATE_SEARCH = 'UPDATE_SEARCH';
 export const SHOW_NOTES = 'SHOW_NOTES';
-export const LOGIN = 'LOGIN';
 export const SIGNING_UP = 'SIGN_UP';
+export const TITLE_SORT = 'TITLE_SORT';
+export const TOGGLE_DELETE = 'TOGGLE_DELETE';
+export const UPDATE_NOTE = 'UPDATE_NOTE';
+export const UPDATE_SEARCH = 'UPDATE_SEARCH';
 export const USER_CREATED = 'USER_CREATED';
-export const ERROR = 'ERROR';
 
 export const addNote = data => ({
   type: ADD_NOTE,
@@ -24,24 +24,17 @@ export const addNote = data => ({
   stamp: data.stamp,
 });
 
-export const updateNote = data => ({
-  type: UPDATE_NOTE,
-  id: data.id,
-  title: data.title,
-  body: data.body,
-});
-
 export const deleteNote = data => ({
   type: DELETE_NOTE,
   id: data.id,
 });
 
-export const toggleDelete = data => ({
-  type: TOGGLE_DELETE,
+export const error = data => ({
+  type: ERROR,
 });
 
-export const titleSort = data => ({
-  type: TITLE_SORT,
+export const login = data => ({
+  type: LOGIN,
 });
 
 export const newestSort = data => ({
@@ -50,11 +43,6 @@ export const newestSort = data => ({
 
 export const oldestSort = data => ({
   type: OLDEST_SORT,
-});
-
-export const updateSearch = data => ({
-  type: UPDATE_SEARCH,
-  input: data.input,
 });
 
 export const showNotes = data => ({
@@ -78,6 +66,26 @@ export const signup = (data) => {
   };
 };
 
-export const login = data => ({
-  type: LOGIN,
+export const titleSort = data => ({
+  type: TITLE_SORT,
+});
+
+export const toggleDelete = data => ({
+  type: TOGGLE_DELETE,
+});
+
+export const updateNote = data => ({
+  type: UPDATE_NOTE,
+  id: data.id,
+  title: data.title,
+  body: data.body,
+});
+
+export const updateSearch = data => ({
+  type: UPDATE_SEARCH,
+  input: data.input,
+});
+
+export const userCreated = data => ({
+  type: USER_CREATED,
 });
