+++ certain-death/src/components/Forms/updateform.js
@@ -2,8 +2,8 @@ import React from 'react';
 import styled from 'styled-components';
 import { connect } from 'react-redux';
 import { Redirect } from 'react-router-dom';
-import { updateNote } from '../actions';
-import Button from './button';
+import { updateNote } from '../../actions';
+import Button from '../Misc/button';
 
 const StyledUpdateForm = styled.div`
   display: flex;
