+++ certain-death/src/components/Rails/homeleftrail.js
@@ -2,9 +2,9 @@ import React from 'react';
 import styled from 'styled-components';
 import { Link } from 'react-router-dom';
 import { connect } from 'react-redux';
-import { titleSort, oldestSort, newestSort } from '../actions';
-import Search from './search';
-import Button from './button';
+import { titleSort, oldestSort, newestSort } from '../../actions';
+import Search from '../Forms/search';
+import Button from '../Misc/button';
 
 const StyledHomeRail = styled.div`
   width: 210px;
